test(profile): Consolidate builtin profiles
